+++ nope/lib/nope/Nope/Platform/Setting/Field.php
@@ -13,17 +13,10 @@ class Field {
   function __construct($id, $properties = []) {
     $this->id = $id;
     $this->properties = (object) $properties;
-    switch($this->properties->type) {
-      default:
-      case 'input':
-        $type = 'input';
-        break;
-      case 'text':
-      case 'model':
-      case 'pair':
-      case 'table':
-        $type = $this->properties->type;
-        break;
+    if(!$this->properties->type) {
+      $type = 'input';
+    } else {
+      $type = $this->properties->type;
     }
     $className = 'Nope\Platform\Setting\Field\\' . S::upperCaseFirst($type);
     $this->instance = new $className($this->id, $this->properties);