+++ nope/lib/register/settings.php
@@ -7,6 +7,7 @@ $nopeSettings = new Setting('nope', [
   'description' => 'General settings',
   'role' => 'admin'
 ]);
+/*
 $nopeSettings->addField(new Setting\Field('headline', [
   'label' => 'Headline',
   'attributes' => [
@@ -22,32 +23,37 @@ $nopeSettings->addField(new Setting\Field('description', [
     'rows' => 5
   ]
 ]));
+*/
 
-/*$nopeSettings->addField(new Setting\Field('table', [
+/*$nopeSettings->addField(new Setting\Field('table1', [
   'label' => 'Table',
   'type' => 'table',
-  'header' => ['A', 'Bb'],
-  'maxRows' => 5
-]));*/
-
-$nopeSettings->addField(new Setting\Field('keyvalue', [
-  'label' => 'Key value',
-  'type' => 'pair'
+  'multiple' => true
 ]));
+*/
 
-/*
+
+/*$nopeSettings->addField(new Setting\Field('keyvalue1', [
+  'label' => 'Key value',
+  'type' => 'pair',
+  'multiple' => true
+]));
+*/
+
+
 $nopeSettings->addField(new Setting\Field('cover', [
   'label' => 'Website cover',
   'type' => 'model',
   'model' => '\Nope\Media',
   'attributes' => [
-    'href' => '#/media',
+    'href' => '#/media?mimetype=image/',
     'multiple' => false,
     'label' => 'Add website cover',
     'preview' => 'icon'
   ]
 ]));
 
+/*
 $testGroup = new Setting\Group('testgroup', [
   'label' => 'Test group!',
   'multiple' => true